+++ src/example/index.ts
@@ -2,3 +2,8 @@ export * from './di'
 export * from './progress'
 export * from './streaming'
 export * from './service'
+export { TodoList } from "./service/TodoList";
+export { TodoApp } from "./service/TodoApp";
+export { TodoSummary } from "./service/TodoSummary";
+export { TodoDetails } from "./service/TodoDetails";
+export { TodoStatus } from "./service/TodoStatus";
